Made input clearer
I very quickly tried Corpus with docs (rather than txt files) and I'm getting gibberish unicode, probably because the docs have super weird formatting. I'll try to sort out how to read in funky doc files while building the scraper this weekend and hopefully we can just throw that to the top of this too. 

If it doesn't work, we can go off the .txt files for now.